+++ collector/collector.go
@@ -212,7 +212,7 @@ func (c *Collector) Collect(ch chan<- prometheus.Metric) {
 		c.ErlangVMMetrics.Collect(ch, report.ErlangVMReport)
 	}
 	if c.EnableSoraClusterMetrics {
-		c.SoraClusterMetrics.Collect(ch, nodeList, report.ClusterReport)
+		c.SoraClusterMetrics.Collect(ch, nodeList, report.ClusterReport, report.ClusterRelay)
 	}
 }