+++ collector/sora_api.go
@@ -8,6 +8,7 @@ type soraGetStatsReport struct {
 	SoraConnectionErrorReport soraConnectionErrorReport `json:"error,omitempty"`
 	ErlangVMReport            erlangVMReport            `json:"erlang_vm,omitempty"`
 	ClusterReport             soraClusterReport         `json:"cluster,omitempty"`
+	ClusterRelay              []soraClusterRelay        `json:"cluster_relay,omitempty"`
 }
 
 type soraConnectionReport struct {
@@ -47,6 +48,7 @@ type soraClientStatistics struct {
 	SoraJsSdk                   int64 `json:"sora_js_sdk"`
 	SoraUnitySdk                int64 `json:"sora_unity_sdk"`
 	ObsStudioWhip               int64 `json:"obs_studio_whip"`
+	ObsStudioWhep               int64 `json:"obs_studio_whep"`
 	SoraPythonSdk               int64 `json:"sora_python_sdk"`
 	Unknown                     int64 `json:"unknown"`
 	WebrtcLoadTestingToolZakuro int64 `json:"webrtc_load_testing_tool_zakuro"`
@@ -140,6 +142,14 @@ type soraClusterNode struct {
 	NodeName        string `json:"node_name"`
 	Mode            string `json:"mode"`
 	Connected       bool   `json:"connected"`
+}
+
+type soraClusterRelay struct {
+	NodeName              string `json:"node_name"`
+	TotalReceivedByteSize int64  `json:"total_received_byte_size"`
+	TotalSentByteSize     int64  `json:"total_sent_byte_size"`
+	TotalReceived         int64  `json:"total_received"`
+	TotalSent             int64  `json:"total_sent"`
 }
 
 type soraLicenseInfo struct {